requirements.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.2.1
anyio==3.3.1
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
click==8.1.3
codecov==2.1.12
colorama==0.4.5
commonmark==0.9.1
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
coverage==6.4.4
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
flake8==5.0.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.21.1
iniconfig==1.1.1
interrogate==1.5.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-language-pack-zh-CN @ http://autodl-public.ks3-cn-beijing.ksyun.com/instance/jupyterlab_language_pack_zh_CN-0.0.1.dev0-py2.py3-none-any.whl
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mmcv==1.6.1
mmdet==2.25.1
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.6
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python-headless==4.6.0.66
packaging==21.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.9.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.5.0
Pygments==2.10.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.1.2
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==12.5.1
rsa==4.7.2
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
supervisor==4.2.2
tabulate==0.8.10
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu111/torch-1.8.1%2Bcu111-cp38-cp38-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu111/torchvision-0.9.1%2Bcu111-cp38-cp38-linux_x86_64.whl
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.1.0
typing-extensions==4.3.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xdoctest==1.0.2
yapf==0.32.0
